// src/autenticacion/dominio/enums/rol-usuario.enum.ts
/**
 * Enumeración que define los roles de usuario disponibles en el sistema
 * Sigue la convención de nomenclatura en español
 */
export enum RolUsuario {
  SUPER_ADMIN = 'SUPER_ADMIN',
  ADMIN = 'ADMIN',
  GERENTE = 'GERENTE',
  SUPERVISOR = 'SUPERVISOR',
  CAJERO = 'CAJERO',
  EDITOR = 'EDITOR',
  VENDEDOR = 'VENDEDOR',
}